pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h5pulsar"
version = "0.0.1"
authors = [
    {name="Rutger van Haasteren", email="rutger@vhaasteren.com"},
]
description = "HDF5 read/write capabilities for Enterprise Pulsar objects"
readme = "README.md"
license = {file="LICENSE"}
requires-python = ">=3.7"
dependencies = [
    "numpy>=1.16.3",
    "astropy>=4.0,!=4.0.1,!=4.0.1.post1",
    "enterprise-pulsar>=3.3.4",
    "jplephem>=2.6",
    "h5py>=3.9.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/vhaasteren/h5pulsar"
"Bug Tracker" = "https://github.com/vhaasteren/h5pulsar/issues"
